+++ lib/extra.dart
@@ -1,0 +1,3 @@
+export 'src/utils/utils.dart';
+export 'src/widgets/clear_focus_background.dart';
+export 'src/widgets/faded_animated_switcher.dart';