+++ lib/src/paged_list_view_fetcher.dart
@@ -1,3 +1,4 @@
+import 'package:fetcher/extra.dart';
 import 'package:fetcher/fetcher.dart';
 import 'package:flutter/material.dart';
 import 'package:value_stream/value_stream.dart';