+++ example/lib/pages/fetch_builder.page.dart
@@ -97,8 +97,8 @@ class _FetchBuilderPageState extends State<FetchBuilderPage> {
             controller: _fetchController1,
             task: fetchTask,
             config: const FetcherConfig(
-              fadeDuration: Duration.zero,    // Disable fade
-              // fadeDuration: Duration(seconds: 2),   // Long fade
+              // fadeDuration: Duration.zero,    // Disable fade
+              fadeDuration: Duration(seconds: 1),   // Long fade
             ),
             builder: (context, data) {
               return Column(