+++ lib/fetcher_bloc.dart
@@ -1,0 +1,6 @@
+export 'fetcher.dart';
+export 'extra.dart';
+export 'package:value_stream/value_stream.dart';
+
+export 'src/bloc/bloc_provider.dart';
+export 'src/bloc/disposable.dart';