+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 3.0.0
-* BREAKING: a bit of renaming & export cleaning.
 * BREAKING: rename `AsyncTaskBuilder` to `SubmitBuilder`.
 * BREAKING: rename `AsyncForm` to `SubmitFormBuilder`.
 * BREAKING: rename `AsyncEditBuilder.commitTask` to `AsyncEditBuilder.submitTask`.
+* BREAKING: all `onSuccess` callbacks are now synchronous.
+* BREAKING: a bit of renaming & export cleaning.
 * New `FetcherConfig.silent` config for cases where loader & error should not be displayed.
 
 ## 2.0.1