+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 4.0.0
+* BREAKING: remove obsolete `clearFocus2` method.
+
 ## 3.1.0
 * Add NewsReaderPage example
 * Add BLoC pattern components, so fetcher can be use with BLoC pattern without any additional dependency.